refactor: centralize reference resolver
